+++ routes/public.php
@@ -5,6 +5,7 @@ use WebBlocks\Cms\Http\Controllers\AdminApi\SiteDomainApiController;
 use WebBlocks\Cms\Http\Controllers\Public\CommentEntryController;
 use WebBlocks\Cms\Http\Controllers\Public\ContactMessageController;
 use WebBlocks\Cms\Http\Controllers\Public\ContentRatingController;
+use WebBlocks\Cms\Http\Controllers\Public\EmbeddedApplicationEntryController;
 use WebBlocks\Cms\Http\Controllers\Public\PackagePublicStatusController;
 use WebBlocks\Cms\Http\Controllers\Public\PageController;
 use WebBlocks\Cms\Http\Controllers\Public\PublicPrivacyConsentController;
@@ -15,6 +16,11 @@ use WebBlocks\Cms\Support\Pages\PagePath;
 use WebBlocks\Cms\WebBlocksCmsServiceProvider;
 
 $publicPageMiddleware = ['web', 'install.required', AddCmsIdentificationHeader::class];
+
+Route::middleware($publicPageMiddleware)
+  ->get('/webblocks-applications/{application}/index.html', EmbeddedApplicationEntryController::class)
+  ->where('application', '[a-z0-9](?:[a-z0-9-]{0,62}[a-z0-9])?')
+  ->name('embedded-applications.entry');
 
 if (config(WebBlocksCmsServiceProvider::PACKAGE_PUBLIC_STATUS_ROUTE_LOADING_CONFIG, false)) {
   Route::middleware($publicPageMiddleware)